Move code from rate_limit to free_usage
